requirements.txt
aioAlice~=1.5.1
aiohttp~=3.8.4
mysqlclient~=2.1.1
psycopg2==2.9.6
python-dotenv~=1.0.0
